Elise animate track

+++ WaveApp/app/src/main/java/se/chalmers/halfwaytospirit/waveapp/Point.java
@@ -4,15 +4,15 @@ package se.chalmers.halfwaytospirit.waveapp;
  * Created by Elise on 07/12/2015.
  */
 public class Point {
-    private int x;
-    private int y;
+    private long x;
+    private long y;
 
     /**
      * Constructor.
      * @param x - the x-coordinate of the point.
      * @param y - the y-coordinate of the point.
      */
-    public Point(int x, int y) {
+    public Point(long x, long y) {
         this.x = x;
         this.y = y;
     }
@@ -21,7 +21,7 @@ public class Point {
      * Gets x.
      * @return the x-coordinate of the point.
      */
-    public int getX() {
+    public long getX() {
         return x;
     }
 
@@ -29,7 +29,7 @@ public class Point {
      * Gets y.
      * @return the y-coordinate of the point.
      */
-    public int getY() {
+    public long getY() {
         return y;
     }
 }
